perf: opt imgui inspector
