+++ web-packages/imgui-demo/src/core/decorators.ts
@@ -1,7 +1,40 @@
+
 export const menuItemStore: Record<string, () => void> = {};
 
 export function menuItem (path: string) {
   return (target: any, key: string, descriptor: PropertyDescriptor) => {
     menuItemStore[path] = descriptor.value;
+  };
+}
+
+export const editorWindowStore: (new () => any)[] = [];
+export function editorWindow () {
+  return (target: any, context?: unknown) => {
+    for (const editorWindowClass of editorWindowStore) {
+      if (editorWindowClass === target) {
+        console.warn(`Class ${target} 重复注册`);
+      }
+    }
+    editorWindowStore.push(target);
+  };
+}
+
+export const editorStore = new Map<Function, new () => any>();
+export function editor (ComponentType: Function) {
+  return (target: any, context?: unknown) => {
+    if (editorStore.get(ComponentType)) {
+      console.warn(`Editor Class ${target} 重复注册`);
+    }
+    editorStore.set(ComponentType, target);
+  };
+}
+
+export const objectInspectorStore = new Map<Function, new () => any>();
+export function objectInspector (ObjectType: Function) {
+  return (target: any, context?: unknown) => {
+    if (objectInspectorStore.get(ObjectType)) {
+      console.warn(`ObjectInspector Class ${target} 重复注册`);
+    }
+    objectInspectorStore.set(ObjectType, target);
   };
 }
