+++ web-packages/imgui-demo/src/custom-editors/editor.ts
@@ -1,0 +1,5 @@
+export class Editor {
+  onInspectorGUI () {
+
+  }
+}
