+++ web-packages/imgui-demo/src/main.ts
@@ -6,6 +6,7 @@ import { ShowDemoWindow } from './imgui/imgui_demo';
 import { MemoryEditor } from './imgui/imgui_memory_editor';
 import './panels';
 import './custom-editors';
+import './object-inspectors';
 import { imGuiIni } from './imgui/imgui-config';
 import { uiManager } from './core/ui-manager';
 import { Project } from './panels';
