+++ web-packages/imgui-demo/src/object-inspectors/index.ts
@@ -1,0 +1,2 @@
+export * from './object-inspectors';
+export * from './vfx-item-inspector';
