+++ web-packages/imgui-demo/src/object-inspectors/object-inspectors.ts
@@ -1,0 +1,7 @@
+export class ObjectInspector {
+  title = '';
+  activeObject: object;
+
+  onGUI () {
+  }
+}
