+++ web-packages/imgui-demo/src/panels/editor-window.ts
@@ -48,15 +48,3 @@ export class EditorWindow {
 
   }
 }
-
-export const editorWindowStore: (new () => EditorWindow)[] = [];
-export function editorWindow () {
-  return (target: any, context?: unknown) => {
-    for (const editorWindowClass of editorWindowStore) {
-      if (editorWindowClass === target) {
-        console.warn(`Class ${target} 重复注册`);
-      }
-    }
-    editorWindowStore.push(target);
-  };
-}
