+++ web-packages/imgui-demo/src/panels/index.ts
@@ -3,4 +3,4 @@ export * from './sequencer';
 export * from './project';
 export * from './main-editor';
 export * from './preview';
-export * from './editor';
+export * from './inspector';
