+++ web-packages/imgui-demo/src/panels/project.ts
@@ -6,8 +6,8 @@ import { GeometryBoxProxy, ModelMeshComponent, Sphere } from '@galacean/effects-
 import { GLTFTools, ModelIO } from '@vvfx/resource-detection';
 import { folderIcon, jsonIcon } from '../asset/images';
 import { AssetDatabase, createPreviewPlayer, generateAssetScene, readFileAsAsData, readFileAsText } from '../core/asset-data-base';
-import { menuItem } from '../core/decorators';
-import { EditorWindow, editorWindow } from './editor-window';
+import { editorWindow, menuItem } from '../core/decorators';
+import { EditorWindow } from './editor-window';
 import { Selection } from '../core/selection';
 import { ImGui, ImGui_Impl } from '../imgui';
 import { FileNode } from '../core/file-node';
