+++ web-packages/imgui-demo/src/panels/sequencer.ts
@@ -1,16 +1,16 @@
 import type { Composition, TrackAsset } from '@galacean/effects';
 import { CompositionComponent, VFXItem } from '@galacean/effects';
-import { menuItem } from '../core/decorators';
+import { editorWindow, menuItem } from '../core/decorators';
 import { GalaceanEffects } from '../ge';
 import { ImGui } from '../imgui';
-import { EditorWindow, editorWindow } from './editor-window';
+import { EditorWindow } from './editor-window';
 
 @editorWindow()
 export class Sequencer extends EditorWindow {
-  isDragging: boolean = false;
-  currentTime: number = 0;
-  trackUIOffset: number = 200;
-  timeCursorPositionX: number = 0;
+  isDragging = false;
+  currentTime = 0;
+  trackUIOffset = 200;
+  timeCursorPositionX = 0;
 
   currentComposition: Composition;
